requirements.txt
sentencepiece
streamlit
transformers
GoogleNews
matplotlib
seaborn
pandas
sacremoses
plotly
python-dotenv
